refactoring_engine/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "refactoring_engine"
version = "0.1.0"
description = "Distributed Codebase Refactoring Engine (Python-only MVP)"
authors = [
    { name = "Your Name", email = "your@email.com" }
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "astor",
    "fastapi",
    "pydantic",
    "gitpython",
    "pytest",
    "openai",
    "uvicorn"
]
